company_portfolio_ai/requirements.txt
streamlit
playwright
beautifulsoup4
python-dotenv
langgraph
langchain-openai
reportlab
tiktoken
